Only valid transactions will be pushed into TransactionPool.

+++ lib/block.go
@@ -59,7 +59,7 @@ func NewBlock(proposer string, round Round, transactions []string, confirmed str
 		Confirmed:    confirmed,
 	}
 
-	log.Debug("NewBlock created", "PrevTotalTxs", round.TotalTxs, "txs", uint64(len(transactions)), "TotalTxs", b.Header.TotalTxs)
+	log.Debug("NewBlock created", "PrevTotalTxs", round.TotalTxs, "txs", len(transactions), "TotalTxs", b.Header.TotalTxs)
 
 	b.Hash = base58.Encode(sebakcommon.MustMakeObjectHash(b))
